// packages/core/src/type.ts
export interface ISeoMetaOptions {
  title: string
  description: string
  url: string
  image?: string
  keywords?: string
  desc?: string
  locale?: string
  siteName?: string
  ignoreTwitter?: Boolean
  twitterUser?: string
  ignoreOG?: Boolean
  defaultImage?: string
  defaultUrl?: string
  defaultDescription?: string
}